+++ test/interpreter/operator-division.spec.ts
@@ -57,16 +57,17 @@ describe('Operator DIVISION', () => {
     expect(engine.getCellValue(adr('A1'))).toEqualError(detailedError(ErrorType.NAME, ErrorMessage.FunctionName('FOOBAR')))
   })
 
-  // Inconsistency with Product 1
   it('range value results in VALUE error', () => {
     const engine = HyperFormula.buildFromArray([
-      ['1', '=10 / A1:A3'],
-      ['9', '=A1:A3 / 10'],
+      ['1'],
+      ['9'],
       ['3'],
-    ])
+      ['=10 / A1:A3'],
+      ['=A1:A3 / 10'],
+    ], {arrays: false})
 
-    expect(engine.getCellValue(adr('B1'))).toEqualError(detailedError(ErrorType.VALUE, ErrorMessage.ScalarExpected))
-    expect(engine.getCellValue(adr('B2'))).toEqualError(detailedError(ErrorType.VALUE, ErrorMessage.ScalarExpected))
+    expect(engine.getCellValue(adr('A4'))).toEqualError(detailedError(ErrorType.VALUE, ErrorMessage.ScalarExpected))
+    expect(engine.getCellValue(adr('A5'))).toEqualError(detailedError(ErrorType.VALUE, ErrorMessage.ScalarExpected))
   })
 
   it('Division propagates errors correctly', () => {