+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added ARRAYFORMULA function. (#630)
 - Added FILTER function. (#668)
 - Added ARRAY_CONSTRAIN function. (#661)
+- Added casting to scalars from non-range arrays. (#663)
+- Added support for range interpolation. (#665)
 
 ### Fixed
 - Fixed an issue with arrays and cruds. (#651)